refactor(gulp): Split up gulpfile to different files
