+++ gulp/tasks/clean.js
@@ -1,0 +1,16 @@
+import del from 'del';
+
+import {mainDirectories} from '../config';
+import {isProdBuild} from '../commandLineArgs';
+
+/**
+ * Delete output directories
+ */
+function clean() {
+	if (isProdBuild()) {
+		return del(mainDirectories.dist);
+	}
+	return del(mainDirectories.dev);
+}
+
+export default clean;
