+++ gulp/tasks/security.js
@@ -1,0 +1,17 @@
+import nsp from 'gulp-nsp';
+import path from 'path';
+
+import {isProdBuild} from '../commandLineArgs';
+
+/**
+ * Check dependencies with help of the node security platform
+ */
+function security(done) {
+	if (isProdBuild()) {
+		nsp({package: path.join(__dirname, '../../package.json')}, done);
+	} else {
+		done();
+	}
+}
+
+export default security;
