+++ package.json
@@ -104,6 +104,10 @@
     ]
   },
   "jest": {
+    "testPathIgnorePatterns": [
+      "<rootDir>/gulp/",
+      "<rootDir>/node_modules/"
+    ],
     "collectCoverage": true,
     "collectCoverageFrom": [
       "src/app/**/*.{js,jsx}",
